extract initializer for actions
